Extensions to Basic Information Cluster; AttrId is now `u32`; report more global elements (Was: Chip tool tests) (#232)
* WIP - run chip-tool tests with rs-matter - TestBasicInformation

* WIP - run chip-tool tests with rs-matter - TestAttributesById

* WIP - run chip-tool tests with rs-matter - TestBasicInformation

* Fix tests; clippy

* Address code review comment

+++ rs-matter/src/data_model/objects/mod.rs
@@ -45,7 +45,7 @@ pub use metadata::*;
 
 pub type EndptId = u16;
 pub type ClusterId = u32;
-pub type AttrId = u16;
+pub type AttrId = u32;
 pub type CmdId = u32;
 
 #[derive(Debug, ToTLV, Copy, Clone)]
